stock_collector/requirements.txt
playwright
pyyaml
pytz
python-dateutil
